+++ Extras/Fixtures/SwiftPackage/Sources/Greeting/Greeting.swift
@@ -1,0 +1,9 @@
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+//  Created by Sam Deane on 24/09/2026.
+//  Copyright © 2026 Elegant Chaos Limited. All rights reserved.
+// -=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-=-
+
+/// Returns a greeting for a name.
+public func greeting(for name: String) -> String {
+  "Hello, \(name)!"
+}